moved the repeated function to helper file
